omplassert kilder og begynt på alle gradientene
